requirements.txt
asgiref==3.2.3
cffi==1.13.2
cryptography==2.8
dj-database-url==0.5.0
Django==3.0.4
django-ckeditor==5.9.0
django-cors-headers==3.3.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-rest-knox==4.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.1.2
psycopg2-binary==2.8.4
pycparser==2.19
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
whitenoise==5.0.1
